// src/types/types.ts
type WeatherResponse = {
  lat: number;
  lon: number;
  timezone: string;
  timezone_offset: number;
  current: {
      dt: number;
      sunrise: number;
      sunset: number;
      temp: number;
      feels_like: number;
      pressure: number;
      humidity: number;
      dew_point: number;
      uvi: number;
      clouds: number;
      visibility: number;
      wind_speed: number;
      wind_deg: number;
      wind_gust: number;
      weather: {
          id: number;
          main: string;
          description: string;
          icon: string;
      }[];
  };
  minutely: {
      dt: number;
      precipitation: number;
  }[];
  hourly: {
      dt: number;
      temp: number;
      feels_like: number;
      pressure: number;
      humidity: number;
      dew_point: number;
      uvi: number;
      clouds: number;
      visibility: number;
      wind_speed: number;
      wind_deg: number;
      wind_gust: number;
      weather: {
          id: number;
          main: string;
          description: string;
          icon: string;
      }[];
      pop: number;
  }[];
  daily: {
      dt: number;
      sunrise: number;
      sunset: number;
      moonrise: number;
      moonset: number;
      moon_phase: number;
      summary: string;
      temp: {
          day: number;
          min: number;
          max: number;
          night: number;
          eve: number;
          morn: number;
      };
      feels_like: {
          day: number;
          night: number;
          eve: number;
          morn: number;
      };
      pressure: number;
      humidity: number;
      dew_point: number;
      wind_speed: number;
      wind_deg: number;
      wind_gust: number;
      weather: {
          id: number;
          main: string;
          description: string;
          icon: string;
      }[];
      clouds: number;
      pop: number;
      rain?: number;
      uvi: number;
  }[];
}

type GeoResponse = {
  name: string;
  local_names: {
    [key: string]: string;
  };
  lat: number;
  lon: number;
  country: string;
  state: string;
}[];

type GetWeatherResponse = {
  status: number;
  message?: string;
  data?: string;
};


export {GeoResponse, GetWeatherResponse, WeatherResponse}